feat: apply next.js dynamic for few bottomsheets
