--- apps/web/src/components/common/floatingButton/FloatingBtn.tsx
@@ -1,12 +1,0 @@
-import { Button } from '@ahhachul/ui'
-import { HTMLAttributes } from 'react'
-import * as S from './styled'
-
-interface FloatingButtonProps extends HTMLAttributes<HTMLButtonElement> {
-  label: string
-  onClick: () => void
-}
-
-export const FloatingButton = ({ label, onClick }: FloatingButtonProps) => (
-  <Button css={S.floatCss} label={label} size="md" type="button" variant="primary" onClick={onClick} />
-)
