--- apps/web/src/components/common/floatingButton/index.ts
@@ -1,1 +1,0 @@
-export * from './FloatingBtn'
