+++ apps/web/src/components/common/index.ts
@@ -2,7 +2,6 @@ export * from './bottomSheet'
 export * from './carousel'
 export * from './drawer'
 export * from './filter'
-export * from './floatingButton'
 export * from './logo'
 export * from './pictureUploader'
 export * from './portal'
