+++ apps/web/src/components/public/floats/FloatingBtn.tsx
@@ -1,0 +1,25 @@
+import { Button } from '@ahhachul/ui'
+import { Theme, css } from '@emotion/react'
+import { HTMLAttributes } from 'react'
+
+interface FloatingButtonProps extends HTMLAttributes<HTMLButtonElement> {
+  label: string
+  onClick: () => void
+}
+
+const FloatingButton = ({ label, onClick }: FloatingButtonProps) => (
+  <Button css={floatCss} label={label} size="md" type="button" variant="primary" onClick={onClick} />
+)
+
+const floatCss = (theme: Theme) => css`
+  position: fixed;
+  bottom: calc(${theme.size.bottomNavbar.height} + 16px);
+  max-width: 132px;
+  right: 20px;
+  padding: 0 24px;
+  z-index: ${theme.zIndex.sticky};
+  filter: drop-shadow(0px 6px 16px rgba(91, 91, 91, 0.25));
+  z-index: ${theme.zIndex.sticky};
+`
+
+export default FloatingButton
